Yay! All tests pass now!
There is one bizarre caveat here:  I found what might be a bug in R in
which if one uses setkey(table, column) and column is a factor, then it
might reorder the rows' keys, which is a disaster.

Check out my note at line 34 of filter_on_max_peptides()
to see the full details.  In any event, when I explicitly did
as.character() on the keys, everything came out as expected; but when I
left the key as a factor (its default), then chaos ensued.

how odd...

The only other caveat: my versions of these functions throw a bunch more
warnings which should be investigated to ensure that no other
shenanigans are happening -- but at least all the tests pass.