created data docstrings and ran roxygen.
In setting up initial docstrings for the data files, I ended up
regenerating a couple of them so that the names match up properly (e.g.:
data(Spyogenes) did not make a Spyogenes object in R, but instead one
named 'data').  I made it create an object named 'Spyogenes' instead and
therefore changed the vignette to take that into account.