I decided to pick off a few R CMD check warnings/notes
I am going to change instances of setkey() to setkeyv() to avoid
annoying NSE semantics and how they are (mis)interpreted by R CMD CHECK.
This suits me just fine, since I am not a fan of having column names
look like variables, I think that is an aspect of NSE which is kind of
nasty.

On the other hand, I am not a fan of having a call to globalVariables()
at the top of the R files, I would rather write a quick comment
mentioning some weirdo variable like '.N' inside the function, then
pre-define it as NULL, then have reshape or data.table or whatever use
it in its default fashion.  This way I have a reminder that we are using
peculiar syntax, the variable is noted, and CMD check is happy.