some minor NSE changes and roxygen hints for ggplot
Added SWATH2stats.R for package level documentation and to have a place
for importing stuff like utils::

This marks the first PR which actually changes anything I think.  It is
_extremely_ mild though, as it only does stuff like use
data.table::setkeyv() and ggplot2::aes_string() in order to step around
the aspects of NSE semantics which annoy R CMD check.

I think before I change anything else, I will write and ask if it is
acceptable that I change anything further.  I would like next to remove
the $ syntax for getting columns or indexes because I think the
things that R does that allow one to type df$rob as a shortcut for
df$robert are begging for hard-to-find bugs.  In my mind, this is a
compelling argument to move all data frames to tbl_df or data.table (but
it is worth noting that I have not done that in my own code, and I
certainly will not do that here unless that is requested).

It appears that when I ran formatr::tidy_dir(), it did not decide on a
single indentation style.  As a result, my editor (emacs) is confused as
to whether to use 2 spaces or 4.  I explicitly told emacs to use 4 as I
think that is what you prefer, but there might still be a couple of
places which are inconsistent.